+++ lib/l10n/app_localizations_ja.dart
@@ -1136,4 +1136,41 @@ class AppLocalizationsJa extends AppLocalizations {
 
   @override
   String get created => 'Created';
+
+  @override
+  String get createMilestone => 'Create Milestone';
+
+  @override
+  String get editMilestone => 'Edit Milestone';
+
+  @override
+  String get deleteMilestone => 'Delete Milestone';
+
+  @override
+  String get deleteMilestoneConfirm =>
+      'Are you sure you want to delete this milestone? This action cannot be undone.';
+
+  @override
+  String get milestoneDeleted => 'Milestone deleted';
+
+  @override
+  String get milestoneTitle => 'Title';
+
+  @override
+  String get milestoneDescription => 'Description';
+
+  @override
+  String get milestoneDescriptionHint => 'Describe this milestone';
+
+  @override
+  String get milestoneDueDate => 'Due Date';
+
+  @override
+  String get milestoneState => 'State';
+
+  @override
+  String get milestoneStateOpen => 'Open';
+
+  @override
+  String get milestoneStateClosed => 'Closed';
 }